OMAP-TEST: TIMER: Fixes for Linux v3.8
For Linux v3.8 the definition OMAP2_DEVICE_TYPE_GP and marcos
cpu_is_xxxx()/soc_is_xxxx() have been moved and this is causing compiler
errors. In v3.8 the cpu_is_xxxx macros for OMAP1 devices are no longer
exposed when building for OMAP2+ devices and the cpu_is_xxxx macros for
OMAP2+ devices are not longer exposed when building for OMAP1 devices.
Workaround this by using the kernel configuration parameter
CONFIG_ARCH_OMAP1 so only the appropriate macros referenced when
building the module. Update the header file where the definitions
and macros are now located.

Signed-off-by: Jon Hunter <jon-hunter@ti.com>