OMAP-TEST: TIMER: DT fix for missing context loss check
When booting with device-tree, the function pointer to the timer context
loss check routine is not initialised. Therefore, when booting with
device-tree, the timer driver has no way to determine if the context has
been lost and needs to be restored. Workaround this for now by keeping
the timer on when in use.

Signed-off-by: Jon Hunter <jon-hunter@ti.com>